Lazy-load font faces.

+++ src/contour/TerminalWidget.h
@@ -56,7 +56,7 @@ class TerminalWidget :
 
     static QSurfaceFormat surfaceFormat();
 
-    int pointsToPixels(int _points) const noexcept;
+    int pointsToPixels(double _points) const noexcept;
 
     void initializeGL() override;
     void resizeGL(int _width, int _height) override;
